requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
schedule==1.2.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.107.0
